bebob-protocols: add protocols crate and move BebobAvc structure
This commit add bebob-protocols crate to implement variety of protocols
specific to each model. At first, move BebobAvc structure.
