+++ libs/bebob/protocols/Cargo.toml
@@ -8,3 +8,8 @@ description = """
 Implementation of protocols defined by BridgeCo. AG and application vendors for DM1000, DM1100, and
 DM1500 ASICs with its BridgeCo. Enhanced Break Out Box (BeBoB) solution.
 """
+
+[dependencies]
+glib = "0.10"
+hinawa = { git = "https://github.com/alsa-project/hinawa-rs.git", tag = "v0.3.0", version = "0.3" }
+ta1394 = { path = "../../ta1394" }
