+++ libs/bebob/runtime/src/maudio/normal_ctls.rs
@@ -12,7 +12,8 @@ use ta1394::{AvcOp, AvcControl};
 use ta1394::general::VendorDependent;
 use ta1394::audio::{AUDIO_SUBUNIT_0_ADDR, CtlAttr, AudioCh, AudioProcessing, ProcessingCtl, AudioFeature, FeatureCtl, AudioSelector};
 
-use super::super::BebobAvc;
+use bebob_protocols::*;
+
 use super::super::model::{IN_METER_NAME, OUT_METER_NAME, OUT_SRC_NAME, OUT_VOL_NAME, HP_SRC_NAME};
 
 use super::common_proto::CommonProto;
