+++ libs/bebob/runtime/src/stanton.rs
@@ -14,7 +14,8 @@ use ta1394::{MUSIC_SUBUNIT_0, Ta1394Avc};
 use ta1394::ccm::{SignalAddr, SignalSubunitAddr};
 use ta1394::audio::{AUDIO_SUBUNIT_0_ADDR, AudioFeature, FeatureCtl, CtlAttr, AudioCh};
 
-use super::BebobAvc;
+use bebob_protocols::*;
+
 use super::common_ctls::ClkCtl;
 use super::model::OUT_VOL_NAME;
